EZP-23471: rename search engine bundles
