--- eZ/Bundle/EzPublishElasticsearchBundle/Resources/config/services.yml
@@ -1,10 +1,0 @@
-parameters:
-    ezpublish.elasticsearch.connection_parameter_factory.class: eZ\Bundle\EzPublishElasticsearchBundle\ApiLoader\ConnectionParameterFactory
-
-services:
-    ezpublish.elasticsearch.connection_parameter_factory:
-        class: %ezpublish.elasticsearch.connection_parameter_factory.class%
-        arguments:
-            - @ezpublish.api.storage_repository_provider
-        calls:
-            - [setContainer, [@service_container]]
