+++ eZ/Bundle/EzPublishElasticsearchSearchEngineBundle/Resources/config/services.yml
@@ -1,0 +1,10 @@
+parameters:
+    ezpublish.elasticsearch.connection_parameter_factory.class: eZ\Bundle\EzPublishElasticsearchSearchEngineBundle\ApiLoader\ConnectionParameterFactory
+
+services:
+    ezpublish.elasticsearch.connection_parameter_factory:
+        class: %ezpublish.elasticsearch.connection_parameter_factory.class%
+        arguments:
+            - @ezpublish.api.storage_repository_provider
+        calls:
+            - [setContainer, [@service_container]]
