--- eZ/Bundle/EzPublishSolrBundle/Resources/config/services.yml
@@ -1,10 +1,0 @@
-parameters:
-    ezpublish.solr.connection_parameter_factory.class: eZ\Bundle\EzPublishSolrBundle\ApiLoader\ConnectionParameterFactory
-
-services:
-    ezpublish.solr.connection_parameter_factory:
-        class: %ezpublish.solr.connection_parameter_factory.class%
-        arguments:
-            - @ezpublish.api.storage_repository_provider
-        calls:
-            - [setContainer, [@service_container]]
